backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Flask API and Database for brewbucks"
authors = ["Nimesh Garg"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^3.0.3"
Flask-SQLAlchemy = "^3.1.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
flask-cors = "^4.0.1"
transformers = "^4.45.2"
pillow = "^11.0.0"
torch = "^2.5.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
